Add instalation tests
