+++ src/json.c
@@ -29,6 +29,11 @@
 #include <nodes/pg_list.h>
 #include <access/htup_details.h>
 
+#include <utils/syscache.h>
+
+PG_MODULE_MAGIC;
+static Oid spectrumOid;
+
 typedef struct
 {
     JsonbIterator   *it;
@@ -286,4 +291,13 @@ Datum load_from_json(PG_FUNCTION_ARGS)
 
     json_ctx_free(json_ctx);
     SRF_RETURN_DONE(funcctx);
+}
+
+void _PG_init()
+{
+#if PG_VERSION_NUM >= 120000
+    spectrumOid = GetSysCacheOid1(TYPENAMENSP, Anum_pg_type_oid, PointerGetDatum("spectrum"));
+#else
+    spectrumOid = GetSysCacheOid1(TYPENAMENSP, PointerGetDatum("spectrum"));
+#endif
 }
