+++ test/expected/base.out
@@ -1,0 +1,29 @@
+\set ECHO none
+1..27
+ok 1 - Type spectrum should exist
+ok 2 - Function load_from_mgf() should exist
+ok 3 - Function load_from_mgf(character varying) should exist
+ok 4 - Function load_from_mgf(oid) should exist
+ok 5 - Function load_from_mgf() should return setof record
+ok 6 - Function spectrum_normalize() should exist
+ok 7 - Function spectrum_normalize(spectrum) should exist
+ok 8 - Function spectrum_normalize() should return spectrum
+ok 9 - Function cosine_greedy() should exist
+ok 10 - Function cosine_greedy(spectrum, spectrum, real, real, real) should exist
+ok 11 - Function cosine_greedy(spectrum, spectrum, real) should exist
+ok 12 - Function cosine_greedy() should return real
+ok 13 - Function spectrum_range() should exist
+ok 14 - Function spectrum_range(spectrum, character varying) should exist
+ok 15 - Function spectrum_range() should return spectrumrange
+ok 16 - Function load_from_json() should exist
+ok 17 - Function load_from_json(jsonb) should exist
+ok 18 - Function load_from_json() should return setof record
+ok 19 - Function cosine_modified() should exist
+ok 20 - Function cosine_modified(spectrum, spectrum, real, real, real, real) should exist
+ok 21 - Function cosine_modified() should return real
+ok 22 - Function intersect_mz() should exist
+ok 23 - Function intersect_mz(spectrum, spectrum, real) should exist
+ok 24 - Function intersect_mz() should return real
+ok 25 - Function precurzor_mz_match() should exist
+ok 26 - Function precurzor_mz_match(real, real, real, character varying) should exist
+ok 27 - Function precurzor_mz_match() should return real
